evaluation function now uses positive and negative integers instead of doubles between 0. and 1.
git-svn-id: file:///Users/Pascal/Development/SVN/Repos/Abalone/trunk@83 1158e906-e6b0-fb4d-a360-84ed95d85be8

+++ AbaloneBase/ComputerPlayerAlphaBeta.h
@@ -28,8 +28,8 @@ public:
   void SetUsedEvaluation(int eval);
 
 protected:
-  double Max(int depth, double alpha, double beta);
-  double Min(int depth, double alpha, double beta);
+  int Max(int depth, int alpha, int beta);
+  int Min(int depth, int alpha, int beta);
 
   GameManager* mySimGameManager;
